+++ tools/level_editor/level_tree_view.vala
@@ -146,7 +146,7 @@ namespace Crown
 				mi = new Gtk.MenuItem.with_label("Rename...");
 				mi.activate.connect(() => {
 					Gtk.Dialog dg = new Gtk.Dialog.with_buttons("New Name"
-						, null
+						, (Gtk.Window)this.get_toplevel()
 						, DialogFlags.MODAL
 						, "Cancel"
 						, ResponseType.CANCEL